Solution to Run Length Encoding.
There are two files in this submission. `run_length_encoding_1.ml` contains
a functional version of encode and an imperative version of decode, while
`run_length_encoding.ml` contains imperative versions of both encode and decode.
The imperative versions run in O(n), where n is the length of the input string.
Although the functional version of encode is also O(n), it makes two passes
in the worst case, making it O(2n).